% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_generated.R
\docType{data}
\name{generated}
\alias{generated}
\title{Generated Data}
\format{
\code{environment} object. You probably only need \code{generated$dt}.
\code{generated} has the following objects:
\itemize{
\item \code{F_cancer} of class(es) "function"
\item \code{F_other} of class(es) "function"
\item \code{F_overall} of class(es) "function"
\item \code{H_cancer} of class(es) "function"
\item \code{H_other} of class(es) "function"
\item \code{H_overall} of class(es) "function"
\item \code{S_cancer} of class(es) "function"
\item \code{S_other} of class(es) "function"
\item \code{S_overall} of class(es) "function"
\item \code{dt} of class(es) "data.table", "data.frame"
\item \code{f_cancer} of class(es) "function"
\item \code{f_other} of class(es) "function"
\item \code{f_overall} of class(es) "function"
\item \code{h_cancer} of class(es) "function"
\item \code{h_other} of class(es) "function"
\item \code{h_overall} of class(es) "function"
\item \code{integrate_data} of class(es) "function"
\item \code{inv_F_cancer} of class(es) "function"
\item \code{inv_F_other} of class(es) "function"
\item \code{inv_F_overall} of class(es) "function"
\item \code{lookup} of class(es) "data.table", "data.frame"
\item \code{lookup_get} of class(es) "function"
\item \code{lookup_get_ages} of class(es) "function"
\item \code{lookup_get_values} of class(es) "function"
\item \code{lookup_stratum_col_nms} of class(es) "function"
\item \code{param_cancer} of class(es) "function"
\item \code{param_other} of class(es) "function"
\item \code{r_cancer} of class(es) "function"
\item \code{r_other} of class(es) "function"
\item \code{r_overall} of class(es) "function"
}
}
\usage{
generated
}
\description{
Data generated for examples and testing.
}
\keyword{datasets}
