% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{level_space_list_to_level_space_data_table}
\alias{level_space_list_to_level_space_data_table}
\title{level_list <- list(
sex = 1:2,
area = 1:5
)
level_space_list_to_level_space_data_table(level_list)}
\usage{
level_space_list_to_level_space_data_table(x)
}
\description{
level_list <- list(
sex = 1:2,
area = data.table::data.table(a1 = c(1, 1, 1, 2, 2), a2 = c(1, 2, 3, 1, 2))
)
level_space_list_to_level_space_data_table(level_list)
}
\details{
level_list <- list(
type = data.table::data.table(t1 = c(1,1,2,2), t2 = c(1,2,3,4)),
area = data.table::data.table(a1 = c(1,1,1,2,2), a2 = c(1,2,3,1,2))
)
level_space_list_to_level_space_data_table(level_list)
}
