% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{stat_count}
\alias{stat_count}
\title{Counts}
\usage{
stat_count(x, by = NULL, subset = NULL, subset_style = c("zeros", "drop")[1])
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

compute counts in this dataset}

\item{subset}{\verb{[NULL, integer, logical]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: no subsetting, i.e. use whole dataset
\item \code{integer}: subset to these rows before computations;
\code{NA} values throw a warning and are not included
\item \code{logical}: subset to rows where this is \code{TRUE} before computing counts;
\code{NA} values throw a warning and are not included; the \code{logical} vector
must have as many elements as the dataset has rows
}}

\item{subset_style}{\verb{[character]} (optional, default \code{"zeros"})
\itemize{
\item \code{"zeros"}: every combination given in \code{joint_column_level_space}
(or existing in data \emph{before} subsetting, if \code{joint_column_level_space}
is \code{NULL}) will be included in output, but strata outside the subset
will have zero counts
\item \code{"drop"}: output will not have strata that are not in the subset
(zero counts are impossible)
}}

\item{stratum_col_nms}{\verb{[NULL, character]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: produces total counts in \code{x}, unless \code{joint_column_level_space}
is not \code{NULL}; then \code{names(joint_column_level_space)} is used and
counts are produced by each of these stratifying column
\item \code{character}: produces counts stratified by these columns
}}

\item{joint_column_level_space}{\verb{[NULL, data.table, list]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: counts are produced using the levels observed in the
dataset; see \code{subset_style} for the effect of subsetting
\item \code{data.table}: describes the stratum column levels and the relationships
between stratum columns; must have a column for all \code{stratum_col_nms}
unless \code{stratum_col_nms} is \code{NULL}
\item \code{list}: a named \code{list}, where each name must be a name of a column in \code{x};
each element of the \code{list} contains the set of levels for the correspoding
column; the \code{list} is turned into a \code{data.table} and that is used as if
you had supplied a \code{data.table}
}}
}
\description{
Produce a stratified count table.
}
\examples{
library("data.table")
sls <- data.table::CJ(sex = 1:2, area_2 = 1:5)
area_sls <- data.table::data.table(
   area_1 = c(1L, 1L, 1L, 2L, 2L), area_2 = 1:5
)
sls <- merge(sls, area_sls, by = "area_2")
data.table::setcolorder(sls, c("sex", "area_1", "area_2"))

my_dataset <- data.table::data.table(
   sex = 1L,
   area_1 = 1L,
   area_2 = 1L
)

stat_count(
  x = my_dataset,
  by = sls
)
stat_count(
  x = my_dataset,
  by = list(sex = 1:2, area_1 = 1:2)
)
stat_count(
  x = my_dataset,
  by = list(sex = 1:2, area = area_sls)
)
}
