% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{stat_prevalent_subject_count}
\alias{stat_prevalent_subject_count}
\title{Prevalence}
\usage{
stat_prevalent_subject_count(
  x,
  follow_up_time_col_nm,
  follow_up_time_window_widths = Inf,
  subject_id_col_nm = NULL,
  by = NULL,
  subset = NULL,
  subset_style = "zero",
  observation_time_points
)
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

dataset containing one or more records by subject}

\item{follow_up_time_col_nm}{\verb{[character]} (mandatory, no default)

name of column in \code{x} which specifies values in the time scale in which
prevalence follow-up windows are determined (e.g. prevalence of subjects
with event within 1, 5, 10 years)}

\item{follow_up_time_window_widths}{\verb{[numeric]} (mandatory, default \code{Inf})

widhts of windowss in time scale given in \code{follow_up_time_col_nm};
e.g. \code{c(1, 5, 10, Inf)} for windows of width 1, 5, 10 and window containing
all subjects}

\item{subject_id_col_nm}{\verb{[character, NULL]} (optional, default \code{NULL})

-\code{character}: name of column in \code{x} which identifies subjects; one subject
may have one or more rows in \code{x}
-\code{NULL}: \code{subject_id_col_nm} is created on-the-fly with the assumption
that each record in \code{x} is a subject}

\item{subset}{\verb{[NULL, integer, logical]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: no subsetting, i.e. use whole dataset
\item \code{integer}: subset to these rows before computations;
\code{NA} values throw a warning and are not included
\item \code{logical}: subset to rows where this is \code{TRUE} before computing counts;
\code{NA} values throw a warning and are not included; the \code{logical} vector
must have as many elements as the dataset has rows
}}

\item{subset_style}{\verb{[character]} (optional, default \code{"zeros"})
\itemize{
\item \code{"zeros"}: every combination given in \code{joint_column_level_space}
(or existing in data \emph{before} subsetting, if \code{joint_column_level_space}
is \code{NULL}) will be included in output, but strata outside the subset
will have zero counts
\item \code{"drop"}: output will not have strata that are not in the subset
(zero counts are impossible)
}}

\item{joint_column_level_space}{\verb{[NULL, data.table, list]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: counts are produced using the levels observed in the
dataset; see \code{subset_style} for the effect of subsetting
\item \code{data.table}: describes the stratum column levels and the relationships
between stratum columns; must have a column for all \code{stratum_col_nms}
unless \code{stratum_col_nms} is \code{NULL}
\item \code{list}: a named \code{list}, where each name must be a name of a column in \code{x};
each element of the \code{list} contains the set of levels for the correspoding
column; the \code{list} is turned into a \code{data.table} and that is used as if
you had supplied a \code{data.table}
}}

\item{stratum_col_nms}{\verb{[character, NULL]} (optional, default \code{NULL})

passed to \link{stat_count}.
\itemize{
\item \code{character}: prevalence counts are stratified by these stratifying columns
\item \code{NULL}: no stratification
}}
}
\description{
Compute subject prevalence at specific time points and optionally
conditioning on time since event at the time points.
}
