% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{stat_rate}
\alias{stat_rate}
\title{Rates}
\usage{
stat_rate(
  x,
  offset_dt,
  stratum_col_nms = NULL,
  adjust_col_nms = NULL,
  subset = NULL,
  adjust_weights = NULL,
  joint_column_level_space = NULL
)
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

passed to \link[basicepistats:stat_count]{basicepistats::stat_count}}

\item{offset_dt}{\verb{[data.table]} (mandatory, no default)

contains offsets (typically mean population sizes) for the counts produced
using \code{x}; see \link[basicepistats:offset_dt]{basicepistats::offset_dt}}

\item{stratum_col_nms}{output will be stratified by these columns}

\item{adjust_col_nms}{output will be adjusted by these columns; see
\link[directadjusting:directly_adjusted_estimates]{directadjusting::directly_adjusted_estimates}}

\item{subset}{passed to \link[basicepistats:stat_count]{basicepistats::stat_count}}

\item{adjust_weights}{passed to
\link[directadjusting:directly_adjusted_estimates]{directadjusting::directly_adjusted_estimates} argument \code{weights}}

\item{joint_column_level_space}{passed to \link[basicepistats:stat_count]{basicepistats::stat_count}}
}
\description{
Event rates table, optionally stratified and directly adjusted.
}
\examples{
library("data.table")
offset_dt <- data.table::CJ(
  sex = 1:2,
  agegroup = 1:18
)
offset_dt[, "offset" := rpois(.N, lambda = 15000)]
}
