% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_table.R
\name{stat_table}
\alias{stat_table}
\alias{stat_table_meta_get}
\alias{stat_table_set}
\alias{as.stat_table}
\alias{as.stat_table.data.table}
\alias{print.stat_table}
\alias{[.stat_table}
\title{Stat Tables}
\usage{
stat_table_meta_get(x)

stat_table_set(x, stratum_col_nms = character(0), value_col_nms = character(0))

stat_table(x, stratum_col_nms = character(0), value_col_nms = character(0))

as.stat_table(x, stratum_col_nms = character(0), value_col_nms = character(0))

\method{as.stat_table}{data.table}(x, stratum_col_nms, value_col_nms)

\method{print}{stat_table}(x, ...)

\method{[}{stat_table}(x, ...)
}
\arguments{
\item{x}{\verb{[R object]} (mandatory, no default)
\itemize{
\item \code{stat_table}, \code{as.stat_table}, \code{stat_table_set}:
an R object to coerce to \code{stat_table}
\item \code{print.stat_table}: object to print
\item \verb{[.stat_table}: object to extract from
\item \code{stat_table_meta_get}: object to extract \code{stat_table} meta data from
}}

\item{stratum_col_nms}{\verb{[character]} (mandatory, no default)

names of stratifying columns in \code{x}}

\item{value_col_nms}{\verb{[character]} (mandatory, no default)

names of columns in \code{x} containing the actual values (statistics)}

\item{...}{passed to next method (see \code{?"["})}
}
\description{
Class \code{stat_table} methods and functions.
}
\examples{

# basicepistats::`[.stat_table`
dt <- data.table::CJ(sex = 1:2, agegroup = 1:18)
dt[, "n" := sample(1e3L, .N)]
basicepistats::stat_table_set(dt, stratum_col_nms = c("sex", "agegroup"),
                              value_col_nms = "n")

stopifnot(
  inherits(dt[1:3, ], class(dt)[1]),
  dt[1:3, sum(.SD[[1]]), .SDcols = "sex"] == 3
)
}
