% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_table_list.R
\name{stat_table_list}
\alias{stat_table_list}
\title{List of Statistics TableS}
\usage{
stat_table_list(varying_arg_list, fixed_arg_list = list(), stat_fun_nm)
}
\arguments{
\item{varying_arg_list}{\verb{[list]} (mandatory, no default)

list of arguments that vary; function named \code{stat_fun_nm} can be called
several times, and each time the arguments supplied here can be different;
each element of this list must in turn be a list and have as name one of
the names of the arguments of the function. each element of
\code{varying_arg_list} must be of the same length. the function will be called
as many times as each element of \code{varying_arg_list} has elements in turn.
e.g. \code{list(by = list("sex", "agegroup", c("sex", "agegroup")))}. see
Examples.}

\item{fixed_arg_list}{\verb{[list]} (optional, default \code{list()})

list of arguments used by function named \code{stat_fun_nm}; these arguments
will be the same each time the function is called. \code{fixed_arg_list}
and \code{varying_arg_list} cannot contain the same arguments.}

\item{stat_fun_nm}{\verb{[character]} (mandatory, no default)

name of a function; must be discoverable by \link{match.fun}}
}
\description{
Call statistics function multiple times to create a list
of statistics tables.
}
\examples{
library("data.table")

my_dataset <- data.table::data.table(
   sex = 1L,
   area_1 = 1L,
   area_2 = 1L
)

stl <- stat_table_list(
  varying_arg_list = list(
    by = list("sex", "area_1")
  ),
  fixed_arg_list = list(x = my_dataset),
  stat_fun_nm = "stat_count"
)

}
