% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{year_based_prevalence}
\alias{year_based_prevalence}
\alias{stat_year_based_prevalent_record_count}
\alias{stat_year_based_prevalent_subject_count}
\alias{stat_year_based_prevalent_subject_count_}
\title{Year-Based Prevalence}
\usage{
stat_year_based_prevalent_record_count(
  x,
  entry_year_col_nm,
  exit_year_col_nm,
  observation_years,
  maximum_follow_up_years = c(1L, 3L, 5L, 1000L),
  by = NULL,
  subset = NULL,
  subset_style = "zeros"
)

stat_year_based_prevalent_subject_count(
  x,
  entry_year_col_nm,
  exit_year_col_nm,
  observation_years,
  subject_id_col_nm,
  maximum_follow_up_years = c(1L, 3L, 5L, 1000L),
  by = NULL,
  subset = NULL,
  subset_style = "zeros"
)

stat_year_based_prevalent_subject_count_(
  x,
  entry_year_col_nm,
  exit_year_col_nm,
  observation_years,
  subject_id_col_nm,
  maximum_follow_up_years = c(1L, 3L, 5L, 1000L),
  by = NULL,
  subset = NULL,
  subset_style = "zeros"
)
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

dataset containing one or more records by subject}

\item{entry_year_col_nm}{\verb{[character]} (mandatory, no default)

name of column in \code{x} for the year of entry into follow-up}

\item{exit_year_col_nm}{\verb{[character]} (mandatory, no default)

name of column in \code{x} for the year of exit from follow-up}

\item{observation_years}{\verb{[observation_years]} (mandatory, no default)

vector of years; the number of prevalent cases / subjects is computed
for the \strong{ends} of these years; e.g. if a person exits follow-up in the
year of observation, this presumably occurs before the last millisecond
of that year, and is \strong{not} considered to be prevalent for that
\code{observation_years} value; see also \strong{Details}}

\item{maximum_follow_up_years}{\verb{[integer]}
(mandatory, default \code{c(1L, 3L, 5L, 1e3L)})

each element of \code{maximum_follow_up_years} defines the upper limit of a window
of follow-up; this is intended to stratify (or filter out) observations
based on how long ago they entered follow-up relative to an individual
point of observation (see \code{observation_years}); each element of this argument
is a exclusive upper limit of an interval (\verb{[a, b[}); e.g. with
\code{observation_years = 2010L} and \code{maximum_follow_up_years = 1L} an observation
is only considered prevalent if it entered follow-up in 2010 (not 2009,
and certainly not 2011); see also \strong{Details}}

\item{by}{\verb{[NULL, data.table, list]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: counts are produced using the levels observed in the
dataset; see \code{subset_style} for the effect of subsetting
\item \code{data.table}: describes the stratum column levels and the relationships
between stratum columns; must have a column for all \code{stratum_col_nms}
unless \code{stratum_col_nms} is \code{NULL}
\item \code{list}: a named \code{list}, where each name must be a name of a column in \code{x};
each element of the \code{list} contains the set of levels for the correspoding
column; the \code{list} is turned into a \code{data.table} and that is used as if
you had supplied a \code{data.table}
}}

\item{subset}{\verb{[NULL, integer, logical]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: no subsetting, i.e. use whole dataset
\item \code{integer}: subset to these rows before computations;
\code{NA} values throw a warning and are not included
\item \code{logical}: subset to rows where this is \code{TRUE} before computing counts;
\code{NA} values throw a warning and are not included; the \code{logical} vector
must have as many elements as the dataset has rows
}}

\item{subset_style}{\verb{[character]} (optional, default \code{"zeros"})
\itemize{
\item \code{"zeros"}: every combination of stratifying columns supplied via \code{by}
(or existing in data \emph{before} subsetting, if \code{by}
is \code{NULL}) will be included in output, but strata outside the subset
will have zero counts
\item \code{"drop"}: output will not have strata that are not in the subset
(zero counts are impossible)
}}

\item{subject_id_col_nm}{\verb{[character]} (mandatory, no default)

name of column in \code{x} which identifies subjects; one subject
may have one or more rows in \code{x}}
}
\description{
Compute prevalent records and subjects using only year-level information.
}
\details{
The following logic determines whether a record is prevalent at a specific
observation year:\preformatted{entry_year <= observation_year < exit_year
}

Additionally, e.g. one-year prevalence is computed by supplying
\code{maximum_follow_up_years = 1L}. To generalise, an observation belongs to
the n-year prevalence group if\preformatted{(observation_year - entry_year) < n
}

Therefore, e.g. an observation is in the one-year prevalence group if
\code{observation_year - entry_year  < 1}, e.g. \code{2020 - 2020 = 0 < 1}.

\itemize{
\item \code{stat_year_based_prevalent_subject_count}: intended for use directly by
the end-user
}

\itemize{
\item \code{stat_year_based_prevalent_subject_count_}: intended for use inside other
functions
}
}
\examples{
library("data.table")
my_dataset <- data.table::data.table(
  id = 1:4,
  sex = c(1L, 1L, 2L, 2L),
  entry_year = 2000:2003,
  exit_year = 2000:2003 + 1:4
)

# NOTE: person that entered in 2003 _is_ counted, person that left in 2003
# is _not_ counted
basicepistats::stat_year_based_prevalent_record_count(
  x = my_dataset,
  entry_year_col_nm = "entry_year",
  exit_year_col_nm = "exit_year",
  maximum_follow_up_years = c(1L, 3L, 5L, 100L),
  by = "sex",
  observation_years = 2003L
)

}
