% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_risk.R
\name{absolute_risk_curve_tools}
\alias{absolute_risk_curve_tools}
\alias{stat_absolute_risk_ag}
\title{Absolute Risk Curve}
\usage{
stat_absolute_risk_ag(
  x,
  stratum_col_nms = NULL,
  timescale_start_col_nm,
  timescale_stop_col_nm,
  event_count_col_nms,
  event_time_amount_col_nms,
  population_time_amount_col_nm
)
}
\arguments{
\item{x}{\verb{[data.table]} (no default)

\code{data.table} with the required columns (see other args). Tt must not have
duplicated rows as identified by
\code{c(stratum_col_nms, timescale_start_col_nm)}.

Chiang, Chin Long, 1986,
Introduction to stochastic processes in biostatistics,
ISBN-13 978-0471155003.}

\item{stratum_col_nms}{\verb{[NULL, character]} (default \code{NULL})
\itemize{
\item \code{NULL}: no stratification of results.
\item \code{character}: Results will be stratified by these columns. E.g.
\code{stratum_col_nms = c("sex", "area")}
}}

\item{timescale_start_col_nm}{\verb{[character]} (no default)

Identifies the start time of the bin for the desired timescale.
E.g. \code{timescale_start_col_nm = "age_group_start"}.

This columns must have one of the classes \code{c("numeric", "integer", "Date")}.
No missing values are allowed.}

\item{timescale_stop_col_nm}{\verb{[character]} (no default)

Identifies the stop time of the bin for the desired timescale.
E.g. \code{timescale_start_col_nm = "age_group_stop"}.

This column must have the same class as \code{x[[timescale_start_col_nm]]}.
No missing values are allowed.}

\item{event_count_col_nms}{\verb{[character]} (no default)

Identifies the columns (one or more) containing counts of events.
E.g. \code{event_count_col_nm = c("event_1", "event_2")}.

Each of these columns must have class \code{integer}. All values must be >= 0.
No missing values are allowed.}

\item{event_time_amount_col_nms}{\verb{[character]} (no default)

Identifies the column (one or more) containing counts of events.
E.g. \code{event_count_col_nm = c("event_1_count", "event_2_count")}.

See section \strong{Under the hood} for more information about the purpose of
this argument.

There must be as many elements in
\code{event_count_col_nms} as in \code{event_time_amount_col_nms}. These two arguments
are assumed to relate to each other elementwise, e.g. the first elements
of both are assumed to belong to the same event.

Each of these columns must be either \code{integer} or \code{numeric}. No missing
values are allowed. All values must be >= 0.}

\item{population_time_amount_col_nm}{\verb{[character]} (no default)

Identifies the column containing time amounts for the (whole, general)
population regardless of what events the subjects in the population may
have encountered. E.g. general population person-years from the national
statistics office. E.g. \code{population_time_amount_col_nm = "pop_pyrs"}.

This column must be either \code{integer} or \code{numeric}. No missing \strong{NOR ZERO}
values are allowed.}
}
\description{
AKA crude risk, sometimes "cumulative incidence function" due to that
estimation method. The tools described here can estimate the curve of the
statistical probability of a subject having an event over a timescale.
E.g. the curve of probabilities of having the first breast cancer diagnosis
can be estimated for some population.
}
\section{Under the hood}{


\code{basicepistats::stat_absolute_risk_ag} implements the method described in
Chiang, Chin Long, 1986,
Introduction to stochastic processes in biostatistics,
ISBN-13 978-0471155003.

The estimation of absolute risk curves is performed via the following steps
on working dataset \code{dt} created based on \code{x}:
\enumerate{
\item At-risk time amounts are computed by subtracting each event time amount
from \code{dt[[population_time_amount_col_nm]]}. This is the total amount of
time in each stratum that the population spent before/without having any
of the events in question. E.g. time the population spent alive and
cancer-free. This is collected into column \code{dt[["at_risk_time_amount"]]}.
\item The hazard rate of each event is estimated via
\code{dt[[event_count_col_nms[i]]] / dt[["at_risk_time_amount"]]} for
each \verb{i in seq_along(event_count_col_nms)}. An "overall hazard rate"
is computed as the sum of all the different hazard rates.
\item The overall survival function is estimated via
\code{exp(-cumsum(dt[["overall_hazard_rate"]]))}. This is performed separately
for each stratum identified by \code{stratum_col_nms}, if any exist.
The "overall event probability" for each bin, conditional on survival
up to the start of the bin, is computed.
\item Event-specific absolute risks are finally estimated via
\code{conditional_overall_risk * h / overall_hazard_rate}, where
\code{h} is an event-specific hazard rate.
}
}

\examples{
dt <- data.table::data.table(
  age_group_start = 0:99,
  age_group_stop = 1:100,
  cancer_count = 1:100,
  death_count = (1:100 + 10L),
  cancer_pyrs = 100:1,
  death_pyrs = rep(0.0, 100),
  pop_pyrs = rep(1e4, 100)
)
dt <- rbind(data.table::copy(dt)[, "sex" := 1L],
            data.table::copy(dt)[, "sex" := 2L])
dt[dt[["sex"]] == 2L, "cancer_count" := rep(1:10, each = 10L)]
dt[dt[["sex"]] == 2L, "cancer_pyrs"  := rep(1.0, 100L)]
ardt <- basicepistats::stat_absolute_risk_ag(
  x = dt,
  stratum_col_nms               = "sex",
  timescale_start_col_nm        = "age_group_start",
  timescale_stop_col_nm         = "age_group_stop",
  event_count_col_nms           = c("cancer_count", "death_count"),
  event_time_amount_col_nms     = c("cancer_pyrs" , "death_pyrs"),
  population_time_amount_col_nm = "pop_pyrs"
)

plot(event_1_absolute_risk ~ age_group_stop,
     data = ardt[ardt[["sex"]] == 1L, ], col = "black", type = "l")
lines(event_1_absolute_risk ~ age_group_stop,
      data = ardt[ardt[["sex"]] == 2L, ], col = "red")
}
