% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_basicepistats.R
\docType{package}
\name{basicepistats}
\alias{basicepistats}
\title{basicepistats}
\description{
Includes functions for computing case counts, prevalence counts,
and other routine epidemiological statistics.
}
\section{News}{

\subsection{News for version 0.2.0}{
\subsection{Object basicepistats::stat_absolute_risk news:}{

\code{basicepistats::stat_absolute_risk} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_count news:}{

\code{basicepistats::stat_count} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_count_ news:}{

\code{basicepistats::stat_count_} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_expr_ news:}{

\code{basicepistats::stat_expr_} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_prevalence_count news:}{

\code{basicepistats::stat_prevalence_count} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_prevalent_record_count news:}{

\code{basicepistats::stat_prevalent_record_count} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_prevalent_subject_count news:}{

\code{basicepistats::stat_prevalent_subject_count} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_prevalent_subject_count_ news:}{

\code{basicepistats::stat_prevalent_subject_count_} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_unique_count news:}{

\code{basicepistats::stat_unique_count} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_unique_count_ news:}{

\code{basicepistats::stat_unique_count_} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_year_based_prevalent_record_count news:}{

\code{basicepistats::stat_year_based_prevalent_record_count} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_year_based_prevalent_subject_count news:}{

\code{basicepistats::stat_year_based_prevalent_subject_count} gains arg
\code{assertion_type}.
}

\subsection{Object basicepistats::stat_year_based_prevalent_subject_count_ news:}{

\code{basicepistats::stat_year_based_prevalent_subject_count_} gains arg
\code{assertion_type}.
}

}
}

