% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_basicepistats.R
\docType{package}
\name{basicepistats}
\alias{basicepistats-package}
\alias{basicepistats}
\title{basicepistats}
\description{
\href{https://github.com/CancerRegistryOfNorway/basicepistats/actions/workflows/R-CMD-check.yaml}{\if{html}{\figure{https://github.com/CancerRegistryOfNorway/basicepistats/actions/workflows/R-CMD-check.yaml/badge.svg}}}

Includes functions for computing case counts, prevalence
counts, and other routine epidemiological statistics.
}
\section{News}{

\subsection{News for version 0.2.4}{
\subsection{basicepistats}{

Replace all \verb{dbc::assert/report_is_data.table*} with
\verb{dbc::assert/report_is_data_table*}. No effect on user.
}

}

\subsection{News for version 0.2.1}{
\subsection{basicepistats::stat_expr_}{

\code{basicepistats::stat_expr_} fix: refer to \code{uniqueN} with
\code{data.table::uniqueN}. Now \code{data.table} need not be loaded with
\code{library(data.table)}.
}

}

\subsection{News for version 0.2.0}{
\subsection{basicepistats::stat_absolute_risk}{

\code{basicepistats::stat_absolute_risk} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_count}{

\code{basicepistats::stat_count} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_count_}{

\code{basicepistats::stat_count_} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_expr_}{

\code{basicepistats::stat_expr_} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_prevalence_count}{

\code{basicepistats::stat_prevalence_count} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_prevalent_record_count}{

\code{basicepistats::stat_prevalent_record_count} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_prevalent_subject_count}{

\code{basicepistats::stat_prevalent_subject_count} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_prevalent_subject_count_}{

\code{basicepistats::stat_prevalent_subject_count_} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_unique_count}{

\code{basicepistats::stat_unique_count} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_unique_count_}{

\code{basicepistats::stat_unique_count_} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_year_based_prevalent_record_count}{

\code{basicepistats::stat_year_based_prevalent_record_count} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_year_based_prevalent_subject_count}{

\code{basicepistats::stat_year_based_prevalent_subject_count} gains arg
\code{assertion_type}.
}

\subsection{basicepistats::stat_year_based_prevalent_subject_count_}{

\code{basicepistats::stat_year_based_prevalent_subject_count_} gains arg
\code{assertion_type}.
}

}
}

\author{
\strong{Maintainer}: Joonas Miettinen \email{joonas.miettinen@cancer.fi}

}
