% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\alias{stat_count}
\alias{stat_count_}
\alias{stat_unique_count}
\alias{stat_unique_count_}
\title{Counts}
\usage{
stat_count(x, by = NULL, subset = NULL, subset_style = c("zeros", "drop")[1])

stat_count_(x, by = NULL, subset = NULL, subset_style = c("zeros", "drop")[1])

stat_unique_count(
  x,
  unique_by,
  by = NULL,
  subset = NULL,
  subset_style = c("zeros", "drop")[1]
)

stat_unique_count_(
  x,
  unique_by,
  by = NULL,
  subset = NULL,
  subset_style = c("zeros", "drop")[1]
)
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

compute counts in this dataset}

\item{by}{\verb{[NULL, data.table, list]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: counts are produced using the levels observed in the
dataset; see \code{subset_style} for the effect of subsetting
\item \code{data.table}: describes the stratum column levels and the relationships
between stratum columns; must have a column for all \code{stratum_col_nms}
unless \code{stratum_col_nms} is \code{NULL}
\item \code{list}: a named \code{list}, where each name must be a name of a column in \code{x};
each element of the \code{list} contains the set of levels for the correspoding
column; the \code{list} is turned into a \code{data.table} and that is used as if
you had supplied a \code{data.table}
}}

\item{subset}{\verb{[NULL, integer, logical]} (optional, default \code{NULL})
\itemize{
\item \code{NULL}: no subsetting, i.e. use whole dataset
\item \code{integer}: subset to these rows before computations;
\code{NA} values throw a warning and are not included
\item \code{logical}: subset to rows where this is \code{TRUE} before computing counts;
\code{NA} values throw a warning and are not included; the \code{logical} vector
must have as many elements as the dataset has rows
}}

\item{subset_style}{\verb{[character]} (optional, default \code{"zeros"})
\itemize{
\item \code{"zeros"}: every combination of stratifying columns supplied via \code{by}
(or existing in data \emph{before} subsetting, if \code{by}
is \code{NULL}) will be included in output, but strata outside the subset
will have zero counts
\item \code{"drop"}: output will not have strata that are not in the subset
(zero counts are impossible)
}}

\item{unique_by}{\verb{[character]} (mandatory, no default)

names of columns in \code{x}; unique combinations of these columns are counted;
e.g. \code{unique_by = "my_subject_id"} to count numbers of subjects by strata}
}
\description{
Produce a stratified count table.
}
\details{
\itemize{
\item \code{stat_count} produces the number of records by strata; this function is
intended to be used directly by the end-user
}

\itemize{
\item \code{stat_count_} produces the number of records by strata; this function is
intended to be used within other functions
}

\itemize{
\item \code{stat_unique_count} produces the number of unique combinations of columns
defined in \code{unique_by}; e.g. the number of unique subjects by strata;
this function is intended to be used directly by the end-user
}

\itemize{
\item \code{stat_unique_count_} produces the number of unique combinations of columns
defined in \code{unique_by}; e.g. the number of unique subjects by strata;
this function is intended to be used inside other functions
}
}
\examples{
library("data.table")
sls <- data.table::CJ(sex = 1:2, area_2 = 1:5)
area_sls <- data.table::data.table(
   area_1 = c(1L, 1L, 1L, 2L, 2L), area_2 = 1:5
)
sls <- merge(sls, area_sls, by = "area_2")
data.table::setcolorder(sls, c("sex", "area_1", "area_2"))

my_dataset <- data.table::data.table(
   sex = 1L,
   area_1 = 1L,
   area_2 = 1L
)

stat_count(
  x = my_dataset,
  by = sls
)
stat_count(
  x = my_dataset,
  by = list(sex = 1:2, area_1 = 1:2)
)
stat_count(
  x = my_dataset,
  by = list(sex = 1:2, area = area_sls)
)
}
