% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{stat_count}
\alias{stat_count}
\title{Counts}
\usage{
stat_count(
  x,
  stratum_col_nms = NULL,
  subset = NULL,
  joint_column_level_space = NULL
)
}
\arguments{
\item{x}{`[data.table]` (mandatory, no default)

compute counts in this dataset}

\item{stratum_col_nms}{`[NULL, character]` (optional, default `NULL`)

- `NULL`: produces total counts in `x`
- `character`: produces counts stratified by these columns}

\item{subset}{`[NULL, integer, logical]` (optional, default `NULL`)

- `NULL`: no subsetting, i.e. use whole dataset
- `integer`: subset to these rows before computing counts;
  `NA` values throw a warning and are not included
- `logical`: subset to rows where this is `TRUE` before computing counts;
  `NA` values throw a warning and are not included}

\item{joint_column_level_space}{`[NULL, data.table]` (optional, default `NULL`)

- `NULL`: counts are produced using the levels observed in the
  (subset of the) dataset
- `data.table`: describes the stratum column levels and the relationships
  between stratum columns; must have a column for all `stratum_col_nms`
  if that was not `NULL`; see **Examples**}
}
\description{
Produce a stratified count table.
}
\examples{
library("data.table")
sls <- data.table::CJ(sex = 1:2, area_2 = 1:5)
area_sls <- data.table::data.table(
   area_1 = c(1L, 1L, 1L, 2L, 2L), area_2 = 1:5
)
sls <- merge(sls, area_sls, by = "area_2")
data.table::setcolorder(sls, c("sex", "area_1", "area_2"))

my_dataset <- data.table::data.table(
   sex = 1L,
   area_1 = 1L,
   area_2 = 1L
)

stat_count(my_dataset, c("sex", "area_1", "area_2"), joint_column_level_space = sls)
}
